Move more logic into PackletImportAnalyzer.ts
