+++ stack/eslint-plugin-packlets/src/mechanics.ts
@@ -5,15 +5,10 @@ import * as path from 'path';
 
 import type { TSESLint, TSESTree } from '@typescript-eslint/experimental-utils';
 import { AST_NODE_TYPES, ESLintUtils } from '@typescript-eslint/experimental-utils';
-import { Path } from './Path';
 
-import { analyze, IResult, MyMessageIds } from './PackletImportAnalyzer';
+import { analyze, analyze2, ILintError, IResult, MyMessageIds, MyMessageIds2 } from './PackletImportAnalyzer';
 
-export type MessageIds =
-  | MyMessageIds
-  | 'bypassed-entry-point'
-  | 'circular-entry-point'
-  | 'packlet-importing-project-file';
+export type MessageIds = MyMessageIds | MyMessageIds2;
 type Options = [];
 
 const mechanics: TSESLint.RuleModule<MessageIds, Options> = {
@@ -94,7 +89,7 @@ const mechanics: TSESLint.RuleModule<MessageIds, Options> = {
         node: TSESTree.ImportDeclaration | TSESTree.ExportNamedDeclaration | TSESTree.ExportAllDeclaration
       ): void => {
         if (node.source?.type === AST_NODE_TYPES.Literal) {
-          if (result.packletsEnabled && result.packletsFolderPath) {
+          if (result.packletsEnabled) {
             // Extract the import/export module path
             // Example: "../../packlets/other-packlet"
             const modulePath = node.source.value;
@@ -111,82 +106,13 @@ const mechanics: TSESLint.RuleModule<MessageIds, Options> = {
               return;
             }
 
-            // Example: /path/to/my-project/src/packlets/my-packlet
-            const inputFileFolder: string = path.dirname(inputFilePath);
-
-            // Example: /path/to/my-project/src/other-packlet/index
-            const importedPath: string = path.resolve(inputFileFolder, modulePath);
-
-            // Is the imported path referring to a file under the src/packlets folder?
-            if (Path.isUnder(importedPath, result.packletsFolderPath)) {
-              // Example: other-packlet/index
-              const importedPathRelativeToPackletsFolder: string = path.relative(
-                result.packletsFolderPath,
-                importedPath
-              );
-              // Example: [ 'other-packlet', 'index' ]
-              const importedPathParts: string[] = importedPathRelativeToPackletsFolder.split(/[\/\\]+/);
-              if (importedPathParts.length > 0) {
-                // Example: 'other-packlet'
-                const importedPackletName: string = importedPathParts[0];
-
-                // We are importing from a packlet. Is the input file part of the same packlet?
-                if (result.packletName && importedPackletName === result.packletName) {
-                  // Yes.  Then our import must NOT use the packlet entry point.
-
-                  // Example: 'index'
-                  //
-                  // We discard the file extension to handle a degenerate case like:
-                  //   import { X } from "../index.js";
-                  const lastPart: string = path.parse(importedPathParts[importedPathParts.length - 1]).name;
-                  let pathToCompare: string;
-                  if (lastPart.toUpperCase() === 'INDEX') {
-                    // Example:
-                    //   importedPath = /path/to/my-project/src/other-packlet/index
-                    //   pathToCompare = /path/to/my-project/src/other-packlet
-                    pathToCompare = path.dirname(importedPath);
-                  } else {
-                    pathToCompare = importedPath;
-                  }
-
-                  // Example: /path/to/my-project/src/other-packlet
-                  const entryPointPath: string = path.join(result.packletsFolderPath, importedPackletName);
-
-                  if (Path.isEqual(pathToCompare, entryPointPath)) {
-                    context.report({
-                      node: node,
-                      messageId: 'circular-entry-point'
-                    });
-                  }
-                } else {
-                  // No.  If we are not part of the same packlet, then the module path must refer
-                  // to the index.ts entry point.
-
-                  // Example: /path/to/my-project/src/other-packlet
-                  const entryPointPath: string = path.join(result.packletsFolderPath, importedPackletName);
-
-                  if (!Path.isEqual(importedPath, entryPointPath)) {
-                    const entryPointModulePath: string = path.posix.relative(
-                      importedPackletName,
-                      inputFileFolder
-                    );
-
-                    context.report({
-                      node: node,
-                      messageId: 'bypassed-entry-point',
-                      data: { entryPointModulePath }
-                    });
-                  }
-                }
-              }
-            } else {
-              // The imported path does NOT refer to a file under the src/packlets folder
-              if (result.packletName) {
-                context.report({
-                  node: node,
-                  messageId: 'packlet-importing-project-file'
-                });
-              }
+            const lint: ILintError | undefined = analyze2(modulePath, result);
+            if (lint) {
+              context.report({
+                node: node,
+                messageId: lint.messageId,
+                data: lint.data
+              });
             }
           }
         }
